requirements.txt
Cerberus>=1.3.4
pandas>=1.4.2
expiringdict>=1.2.1
marshmallow>=3.15.0
requests>=2.27.1
cryptocmd>=0.6.1
Unidecode>=1.3.4
pytz>=2022.1
lxml>=4.8.0
setuptools>=60.2.0
yfinance>=0.1.70
numpy>=1.22.4
punq>=0.6.2
PyJWT>=2.3.0
python-dotenv>=0.20.0
fastapi>=0.77.1
starlette>=0.19.1
kink>=0.6.3
motor>=3.0.0
pymongo>=4.1.1
dnspython>=2.2.1
cython>=0.29.28
colorama>=0.4.4
celery >=5.2.6